Added search function.
Refactored photo shouts controller, left text shouts unchanged for comparison.
Added a hashtag feature so text shouts could be searched for matching hashtags.
Added sunspot gems. This overrides the hashtag search but allows for images to be searched too. Index method added to text and photo shouts for the searchable function from solr.

One issue remaining. To search photos the file extension has to be included. This is undesired.